backend-notificationservice/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
celery==5.4.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
Django==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cockroachdb==5.0
django-cors-headers==4.4.0
django-fcm==0.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
filelock==3.15.4
firebase-admin==6.5.0
frozenlist==1.4.1
future==1.0.0
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
redis==5.0.8
sqlparse==0.5.1
twilio==9.2.3
tzdata==2024.1
urllib3==2.2.2
pika==1.2.1
